Update index.php
The file functions.php is now unique for each theme and is moved from the root of your WonderCMS installation, to your theme folder in /themes/yourTheme/functions.php.

Moving this in a way described will also enable adding custom editable areas for each theme. This is very good news for theme developers.